+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # CHANGELOG
 
-## master
+## 6.0.3 Sep 19, 2021
+
+Upgrade priority: Low. Recommended for all users on 6.0.2 using `api.query.*.*.at` calls
 
 Changes:
 
+- Fix `api.query.<module>.<method>.at` queries
 - Adjust typegen formatting for metadata documentation